+++ gulp/sass.js
@@ -5,19 +5,19 @@
   also includes sourcemaps
 */
 
-var gulp = require('gulp')
-var sass = require('gulp-sass')
-var sourcemaps = require('gulp-sourcemaps')
+const gulp = require('gulp')
+const sass = require('gulp-sass')
+const sourcemaps = require('gulp-sourcemaps')
 
-var config = require('./config.json')
+const config = require('./config.json')
 
 gulp.task('sass', function () {
   return gulp.src(config.paths.assets + '/sass/*.scss')
   .pipe(sourcemaps.init())
   .pipe(sass({outputStyle: 'expanded',
-    includePaths: ['govuk_modules/govuk_frontend_toolkit/stylesheets',
-      'govuk_modules/govuk_template/assets/stylesheets',
-      'govuk_modules/govuk-elements-sass/']}).on('error', sass.logError))
+    includePaths: [/* 'govuk_modules/govuk_frontend_toolkit/stylesheets', */
+      'node_modules/govuk_template_jinja/assets/stylesheets',
+      'node_modules']}).on('error', sass.logError))
   .pipe(sourcemaps.write())
   .pipe(gulp.dest(config.paths.public + '/stylesheets/'))
 })
@@ -26,9 +26,9 @@ gulp.task('sass-documentation', function () {
   return gulp.src(config.paths.docsAssets + '/sass/*.scss')
   .pipe(sourcemaps.init())
   .pipe(sass({outputStyle: 'expanded',
-    includePaths: ['govuk_modules/govuk_frontend_toolkit/stylesheets',
-      'govuk_modules/govuk_template/assets/stylesheets',
-      'govuk_modules/govuk-elements-sass/']}).on('error', sass.logError))
+    includePaths: [/* 'govuk_modules/govuk_frontend_toolkit/stylesheets', */
+      'node_modules/govuk_template_jinja/assets/stylesheets',
+      'node_modules']}).on('error', sass.logError))
   .pipe(sourcemaps.write())
   .pipe(gulp.dest(config.paths.public + '/stylesheets/'))
 })