+++ server.js
@@ -1,21 +1,28 @@
-require('dotenv').config()
-var crypto = require('crypto')
-var path = require('path')
-var express = require('express')
-var session = require('express-session')
-var nunjucks = require('nunjucks')
-var routes = require('./app/routes.js')
-var documentationRoutes = require('./docs/documentation_routes.js')
-var favicon = require('serve-favicon')
-var app = express()
-var documentationApp = express()
-var bodyParser = require('body-parser')
-var browserSync = require('browser-sync')
-var config = require('./app/config.js')
-var utils = require('./lib/utils.js')
-var packageJson = require('./package.json')
+// Core dependencies
+const crypto = require('crypto')
+const path = require('path')
 
-// Grab environment variables specified in Procfile or as Heroku config vars
+// NPM dependencies
+const bodyParser = require('body-parser')
+const browserSync = require('browser-sync')
+const dotenv = require('dotenv')
+const express = require('express')
+const favicon = require('serve-favicon')
+const nunjucks = require('nunjucks')
+const session = require('express-session')
+
+// Local dependencies
+const config = require('./app/config.js')
+const documentationRoutes = require('./docs/documentation_routes.js')
+const packageJson = require('./package.json')
+const routes = require('./app/routes.js')
+const utils = require('./lib/utils.js')
+
+const app = express()
+const documentationApp = express()
+dotenv.config()
+
+// Set up configuration variables
 var releaseVersion = packageJson.version
 var username = process.env.USERNAME
 var password = process.env.PASSWORD
@@ -25,6 +32,7 @@ var useAutoStoreData = process.env.USE_AUTO_STORE_DATA || config.useAutoStoreDat
 var useHttps = process.env.USE_HTTPS || config.useHttps
 var useBrowserSync = config.useBrowserSync
 var analyticsId = process.env.ANALYTICS_TRACKING_ID
+var gtmId = process.env.GOOGLE_TAG_MANAGER_TRACKING_ID
 
 env = env.toLowerCase()
 useAuth = useAuth.toLowerCase()
@@ -40,25 +48,24 @@ promoMode = promoMode.toLowerCase()
 // Disable promo mode if docs aren't enabled
 if (!useDocumentation) promoMode = 'false'
 
-// Force HTTPs on production connections. Do this before asking for basicAuth to
-// avoid making users fill in the username/password twice (once for `http`, and
-// once for `https`).
-
+// Force HTTPS on production. Do this before using basicAuth to avoid
+// asking for username/password twice (for `http`, then `https`).
 var isSecure = (env === 'production' && useHttps === 'true')
-
 if (isSecure) {
   app.use(utils.forceHttps)
   app.set('trust proxy', 1) // needed for secure cookies on heroku
 }
 
-// Authenticate against the environment-provided credentials, if running
-// the app in production (Heroku, effectively)
+// Ask for username and password on production
 if (env === 'production' && useAuth === 'true') {
   app.use(utils.basicAuth(username, password))
 }
 
 // Set up App
-var appViews = [path.join(__dirname, '/app/views/'), path.join(__dirname, '/lib/')]
+var appViews = [path.join(__dirname, '/app/views/'),
+  path.join(__dirname, '/lib/'),
+  path.join(__dirname, '/node_modules/govuk_template_jinja/views/layouts'),
+  path.join(__dirname, '/node_modules/@govuk-frontend/')]
 
 var nunjucksAppEnv = nunjucks.configure(appViews, {
   autoescape: true,
@@ -67,26 +74,28 @@ var nunjucksAppEnv = nunjucks.configure(appViews, {
   watch: true
 })
 
-// Nunjucks filters
+// Add Nunjucks filters
 utils.addNunjucksFilters(nunjucksAppEnv)
 
 // Set views engine
 app.set('view engine', 'html')
 
 // Middleware to serve static assets
+app.use('/icons', express.static(path.join(__dirname, '/node_modules/@govuk-frontend/icons')))
 app.use('/public', express.static(path.join(__dirname, '/public')))
-app.use('/public', express.static(path.join(__dirname, '/govuk_modules/govuk_template/assets')))
-app.use('/public', express.static(path.join(__dirname, '/govuk_modules/govuk_frontend_toolkit')))
-app.use('/public/images/icons', express.static(path.join(__dirname, '/govuk_modules/govuk_frontend_toolkit/images')))
-app.use('/public/codemirror', express.static(path.join(__dirname, '/node_modules/codemirror')))
-app.use('/public/highlight-within-textarea', express.static(path.join(__dirname, '/node_modules/highlight-within-textarea')))
+app.use('/public', express.static(path.join(__dirname, '/node_modules/govuk_template_jinja/assets')))
+// app.use('/public', express.static(path.join(__dirname, '/node_modules/govuk_frontend_toolkit')))
+// app.use('/public/images/icons', express.static(path.join(__dirname, '/node_modules/govuk_frontend_toolkit/images')))
 
 // Elements refers to icon folder instead of images folder
-app.use(favicon(path.join(__dirname, 'govuk_modules', 'govuk_template', 'assets', 'images', 'favicon.ico')))
+app.use(favicon(path.join(__dirname, 'node_modules', 'govuk_template_jinja', 'assets', 'images', 'favicon.ico')))
 
 // Set up documentation app
 if (useDocumentation) {
-  var documentationViews = [path.join(__dirname, '/docs/views/'), path.join(__dirname, '/lib/')]
+  var documentationViews = [path.join(__dirname, '/docs/views/'),
+    path.join(__dirname, '/lib/'),
+    path.join(__dirname, '/node_modules/govuk_template_jinja/views/layouts'),
+    path.join(__dirname, '/node_modules/@govuk-frontend/')]
 
   var nunjucksDocumentationEnv = nunjucks.configure(documentationViews, {
     autoescape: true,
@@ -109,6 +118,7 @@ app.use(bodyParser.urlencoded({
 
 // Add variables that are available in all views
 app.locals.analyticsId = analyticsId
+app.locals.gtmId = gtmId
 app.locals.asset_path = '/public/'
 app.locals.useAutoStoreData = (useAutoStoreData === 'true')
 app.locals.cookieText = config.cookieText
@@ -129,74 +139,14 @@ app.use(session({
   secret: crypto.randomBytes(64).toString('hex')
 }))
 
-// add nunjucks function called 'checked' to populate radios and checkboxes,
-// needs to be here as it needs access to req.session and nunjucks environment
-var addCheckedFunction = function (app, nunjucksEnv) {
-  app.use(function (req, res, next) {
-    nunjucksEnv.addGlobal('checked', function (name, value) {
-      // check session data exists
-      if (req.session.data === undefined) {
-        return ''
-      }
-
-      var storedValue = req.session.data[name]
-
-      // check the requested data exists
-      if (storedValue === undefined) {
-        return ''
-      }
-
-      var checked = ''
-
-      // if data is an array, check it exists in the array
-      if (Array.isArray(storedValue)) {
-        if (storedValue.indexOf(value) !== -1) {
-          checked = 'checked'
-        }
-      } else {
-        // the data is just a simple value, check it matches
-        if (storedValue === value) {
-          checked = 'checked'
-        }
-      }
-      return checked
-    })
-
-    next()
-  })
-}
-
+// Automatically store all data users enter
 if (useAutoStoreData === 'true') {
   app.use(utils.autoStoreData)
-  addCheckedFunction(app, nunjucksAppEnv)
-  addCheckedFunction(documentationApp, nunjucksDocumentationEnv)
+  utils.addCheckedFunction(nunjucksAppEnv)
+  utils.addCheckedFunction(nunjucksDocumentationEnv)
 }
 
-app.use(function (req, res, next) {
-  nunjucksAppEnv.addGlobal('referrer', function (name, value) {
-    var referrer = req.get('Referrer');
-
-    if (referrer) {
-      referrer = referrer.split('/').pop();
-    }
-
-    return referrer;
-  })
-  next()
-})
-
-// Disallow search index idexing
-app.use(function (req, res, next) {
-  // Setting headers stops pages being indexed even if indexed pages link to them.
-  res.setHeader('X-Robots-Tag', 'noindex')
-  next()
-})
-
-app.get('/robots.txt', function (req, res) {
-  res.type('text/plain')
-  res.send('User-agent: *\nDisallow: /')
-})
-
+// Clear all data in session if you open /prototype-admin/clear-data
 app.get('/prototype-admin/clear-data', function (req, res) {
   req.session.destroy()
   res.render('prototype-admin/clear-data')
@@ -210,13 +160,13 @@ if (promoMode === 'true') {
     res.redirect('/docs')
   })
 
-  // allow search engines to index the prototype kit promo site
+  // Allow search engines to index the prototype kit promo site
   app.get('/robots.txt', function (req, res) {
     res.type('text/plain')
     res.send('User-agent: *\nAllow: /')
   })
 } else {
-  // Disallow search index idexing
+  // Prevent search indexing
   app.use(function (req, res, next) {
     // Setting headers stops pages being indexed even if indexed pages link to them.
     res.setHeader('X-Robots-Tag', 'noindex')
@@ -229,7 +179,7 @@ if (promoMode === 'true') {
   })
 }
 
-// routes (found in app/routes.js)
+// Load routes (found in app/routes.js)
 if (typeof (routes) !== 'function') {
   console.log(routes.bind)
   console.log('Warning: the use of bind in routes is deprecated - please check the prototype kit documentation for writing routes.')
@@ -238,7 +188,7 @@ if (typeof (routes) !== 'function') {
   app.use('/', routes)
 }
 
-// Returns a url to the zip of the latest release on github
+// Redirect to the zip of the latest release of the prototype kit on GitHub
 app.get('/prototype-admin/download-latest', function (req, res) {
   var url = utils.getLatestRelease()
   res.redirect(url)
@@ -280,16 +230,15 @@ if (useDocumentation) {
   })
 }
 
-// redirect all POSTs to GETs - this allows users to use POST for autoStoreData
+// Redirect all POSTs to GETs - this allows users to use POST for autoStoreData
 app.post(/^\/([^.]+)$/, function (req, res) {
   res.redirect('/' + req.params[0])
 })
 
 console.log('\nGOV.UK Prototype kit v' + releaseVersion)
-// Display warning not to use kit for production services.
 console.log('\nNOTICE: the kit is for building prototypes, do not use it for production services.')
 
-// start the app
+// Find a free port and start the server
 utils.findAvailablePort(app, function (port) {
   console.log('Listening on port ' + port + '   url: http://localhost:' + port)
   if (env === 'production' || useBrowserSync === 'false') {